refactor: integrate ParticleTests in correct package
